Separate logic for matrices and vectors
